chore: #2 make project portable
- Create a single initial migration script
- Remove historical migration scripts in favor of the new one
- Update sql template env var names
