--- app/database/alembic/versions/2025_12_05_1419-043051aed244_update_user_table.py
@@ -1,82 +1,0 @@
-"""Update user table.
-
-Revision ID: 043051aed244
-Revises: a260d3799d2c
-Create Date: 2025-12-05 14:19:26.427671
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "043051aed244"
-down_revision: str | Sequence[str] | None = "a260d3799d2c"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column("users", sa.Column("name", sa.String(length=100), nullable=True))
-    op.add_column("users", sa.Column("last_name", sa.String(length=100), nullable=True))
-    op.add_column(
-        "users",
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=False,
-        ),
-    )
-    op.add_column(
-        "users",
-        sa.Column(
-            "modified_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=False,
-        ),
-    )
-    op.alter_column(
-        "users",
-        "email",
-        existing_type=sa.VARCHAR(),
-        type_=sa.String(length=255),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "password",
-        existing_type=sa.VARCHAR(),
-        type_=sa.String(length=255),
-        existing_nullable=False,
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column(
-        "users",
-        "password",
-        existing_type=sa.VARCHAR(length=255),
-        type_=sa.VARCHAR(),
-        existing_nullable=False,
-    )
-    op.alter_column(
-        "users",
-        "email",
-        existing_type=sa.VARCHAR(length=255),
-        type_=sa.VARCHAR(),
-        existing_nullable=False,
-    )
-    op.drop_column("users", "modified_at")
-    op.drop_column("users", "created_at")
-    op.drop_column("users", "last_name")
-    op.drop_column("users", "name")
-    # ### end Alembic commands ###
