--- app/database/alembic/versions/2025_12_14_1104-6c70063fa654_add_roles_table.py
@@ -1,55 +1,0 @@
-"""Add roles table.
-
-Revision ID: 6c70063fa654
-Revises: eab06c628d12
-Create Date: 2025-12-14 11:04:36.585638
-
-"""
-
-from collections.abc import Sequence
-
-import sqlalchemy as sa
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "6c70063fa654"
-down_revision: str | Sequence[str] | None = "eab06c628d12"
-branch_labels: str | Sequence[str] | None = None
-depends_on: str | Sequence[str] | None = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "roles",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sa.String(length=255), nullable=False),
-        sa.Column(
-            "created_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=False,
-        ),
-        sa.Column(
-            "modified_at",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=False,
-        ),
-        sa.PrimaryKeyConstraint("id", name=op.f("pk_roles")),
-        sa.UniqueConstraint("name", name=op.f("uq_roles_name")),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-    )
-    op.add_column("users", sa.Column("role_id", sa.Integer(), nullable=False))
-    op.create_foreign_key(op.f("fk_users_role_id_roles"), "users", "roles", ["role_id"], ["id"])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(op.f("fk_users_role_id_roles"), "users", type_="foreignkey")
-    op.drop_column("users", "role_id")
-    op.drop_table("roles")
-    # ### end Alembic commands ###
